backend/requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
python-dotenv
apscheduler
transformers
torch
deep-translator
langdetect
vaderSentiment
numpy
google-generativeai
